@@ -19,31 +19,33 @@
         }
     ],
     "require": {
-        "php": "^7.3 || ^8",
+        "php": "^7.4 || ^8.0",
         "symfony/polyfill-php81": "^1.23"
     },
     "require-dev": {
-        "captainhook/captainhook": "^5.3",
-        "dealerdirect/phpcodesniffer-composer-installer": "^0.7.0",
-        "ergebnis/composer-normalize": "^2.6",
-        "fakerphp/faker": "^1.5",
-        "hamcrest/hamcrest-php": "^2",
-        "jangregor/phpstan-prophecy": "^0.8",
-        "mockery/mockery": "^1.3",
+        "captainhook/plugin-composer": "^5.3",
+        "ergebnis/composer-normalize": "^2.28.3",
+        "fakerphp/faker": "^1.21",
+        "hamcrest/hamcrest-php": "^2.0",
+        "jangregor/phpstan-prophecy": "^1.0",
+        "mockery/mockery": "^1.5",
+        "php-parallel-lint/php-console-highlighter": "^1.0",
+        "php-parallel-lint/php-parallel-lint": "^1.3",
+        "phpcsstandards/phpcsutils": "^1.0.0-rc1",
         "phpspec/prophecy-phpunit": "^2.0",
-        "phpstan/extension-installer": "^1",
-        "phpstan/phpstan": "^0.12.32",
-        "phpstan/phpstan-mockery": "^0.12.5",
-        "phpstan/phpstan-phpunit": "^0.12.11",
-        "phpunit/phpunit": "^8.5 || ^9",
-        "psy/psysh": "^0.10.4",
-        "slevomat/coding-standard": "^6.3",
-        "squizlabs/php_codesniffer": "^3.5",
-        "vimeo/psalm": "^4.4"
+        "phpstan/extension-installer": "^1.2",
+        "phpstan/phpstan": "^1.9",
+        "phpstan/phpstan-mockery": "^1.1",
+        "phpstan/phpstan-phpunit": "^1.3",
+        "phpunit/phpunit": "^9.5",
+        "psalm/plugin-mockery": "^1.1",
+        "psalm/plugin-phpunit": "^0.18.4",
+        "ramsey/coding-standard": "^2.0.3",
+        "ramsey/conventional-commits": "^1.3",
+        "vimeo/psalm": "^5.4"
     },
-    "config": {
-        "sort-packages": true
-    },
+    "minimum-stability": "RC",
+    "prefer-stable": true,
     "autoload": {
         "psr-4": {
             "Ramsey\\Collection\\": "src/"
