@@ -80,10 +80,8 @@ class Collection extends AbstractCollection
      *
      * A collection's type is immutable once it is set. For this reason, this
      * property is set private.
-     *
-     * @var string
      */
-    private $collectionType;
+    private string $collectionType;
 
     /**
      * Constructs a collection object of the specified type, optionally with the
