@@ -14,9 +14,11 @@ declare(strict_types=1);
 
 namespace Ramsey\Collection\Exception;
 
+use RuntimeException;
+
 /**
  * Thrown when attempting to operate on collections of differing types.
  */
-class CollectionMismatchException extends \RuntimeException
+class CollectionMismatchException extends RuntimeException
 {
 }
