@@ -14,9 +14,11 @@ declare(strict_types=1);
 
 namespace Ramsey\Collection\Exception;
 
+use RuntimeException;
+
 /**
  * Thrown when attempting to access an element that does not exist.
  */
-class NoSuchElementException extends \RuntimeException
+class NoSuchElementException extends RuntimeException
 {
 }
